Fix some mixed signed/unsigned operations.

+++ qesource/plot2d/plot.cc
@@ -134,7 +134,7 @@ void* readdata(void *x)
 	pthread_mutex_lock(&M);
 	swap(CE,E);
 	pthread_mutex_unlock(&M);
-	for(int i = 0; i < E.size(); i++) delete E[i];
+	for(size_t i = 0; i < E.size(); i++) delete E[i];
 	E.clear(); }
       else if (c == 'E') { /******* Exit! *******************/
 	return 0;
@@ -155,7 +155,7 @@ void display()
 {
   pthread_mutex_lock(&M);
   glClear(GL_COLOR_BUFFER_BIT);
-  for(int i = 0; i < CE.size(); i++)
+  for(size_t i = 0; i < CE.size(); i++)
     CE[i]->glRend(Colors);
   pthread_mutex_unlock(&M);  
   glutSwapBuffers();
@@ -286,7 +286,7 @@ void SNoverSR::glRend(const CADColors &C)
   // 128 or so points that can appear.
   C.glSetColor(colorType,'D');
   glBegin(GL_LINE_STRIP);
-  for(int i = 0; i < V.size(); i++)
+  for(size_t i = 0; i < V.size(); i++)
     glVertex2(V[i]);
   glEnd();
 }
@@ -328,7 +328,7 @@ bool SRoverSR::read(istream &in)
     else 
     {
       cerr << "Sector over sector in unknown format!" << endl;
-      for(int i = 0; i < V.size(); i++)
+      for(size_t i = 0; i < V.size(); i++)
 	cerr << V[i] << endl;
       exit(1);
     }
